refactor(projects): store projectAccess itself in the shared memo
One struct instead of a stored entry plus a wrapper: the type is
unexported, so exported fields cost nothing. The cache helpers move
next to the resolver that uses them.

Addresses pr-swarm findings: duplicate value type, stale memo comment,
undocumented bare-bean sentinel.
