--- pkg/models/project_access_cache.go
@@ -1,45 +1,0 @@
-// Vikunja is a to-do list application to facilitate your life.
-// Copyright 2018-present Vikunja and contributors. All rights reserved.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-package models
-
-import (
-	"strconv"
-	"time"
-
-	"code.vikunja.io/api/pkg/db"
-)
-
-const (
-	projectAccessCacheTTL       = 30 * time.Second
-	projectAccessCacheKeyPrefix = "project-access-"
-)
-
-func projectAccessCacheKey(userID int64) string {
-	return projectAccessCacheKeyPrefix + strconv.FormatInt(userID, 10)
-}
-
-func invalidateProjectAccess(userID int64) {
-	if userID == 0 {
-		invalidateAllProjectAccess()
-		return
-	}
-	db.InvalidateShared(projectAccessCacheKey(userID))
-}
-
-func invalidateAllProjectAccess() {
-	db.InvalidateSharedPrefix(projectAccessCacheKeyPrefix)
-}
